# docs/api-reference/openapi.json
{
  "info": {
    "description": "Generated from the live Flask url_map \u2014 see potato/server_utils/openapi_spec.py.\n\nMost endpoints are session-authenticated: POST /auth with `username` and `password` as form data, then reuse the session cookie. `x-potato-auth` records the decorators guarding an operation, and `x-potato-requires-config` marks operations that only exist when the named config is enabled.\n\nResponse schemas are intentionally not modelled; this spec is an authoritative index of what exists, not a payload contract.",
    "title": "Potato HTTP API",
    "version": "2.8.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/behavioral_data/{instance_id}": {
      "get": {
        "operationId": "get_get_behavioral_data",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "summary": "Get behavioral data for a specific instance.",
        "tags": [
          "core"
        ],
        "x-potato-auth": [
          "inline:session_username"
        ]
      }
    },
...
    "/api/mcp/manifest": {
      "get": {
        "operationId": "get_mcp_manifest",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "summary": "What this task exposes. Requires a token but no particular tool grant.",
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/add_items": {
      "post": {
        "operationId": "post_mcp_tool_add_items",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/assign_items": {
      "post": {
        "operationId": "post_mcp_tool_assign_items",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/delete_annotations": {
      "post": {
        "operationId": "post_mcp_tool_delete_annotations",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/export_data": {
      "post": {
        "operationId": "post_mcp_tool_export_data",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/get_agreement": {
      "post": {
        "operationId": "post_mcp_tool_get_agreement",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/get_config": {
      "post": {
        "operationId": "post_mcp_tool_get_config",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/get_item": {
      "post": {
        "operationId": "post_mcp_tool_get_item",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/get_progress": {
      "post": {
        "operationId": "post_mcp_tool_get_progress",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/get_status": {
      "post": {
        "operationId": "post_mcp_tool_get_status",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/list_annotators": {
      "post": {
        "operationId": "post_mcp_tool_list_annotators",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/list_items": {
      "post": {
        "operationId": "post_mcp_tool_list_items",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
    "/api/mcp/tools/submit_annotation": {
      "post": {
        "operationId": "post_mcp_tool_submit_annotation",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "tags": [
          "mcp"
        ],
        "x-potato-auth": [
          "mcp_tool",
          "inline:agent_token"
        ],
        "x-potato-requires-config": "mcp.enabled: true, and every tool named in mcp.tools"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}